made derivable dependency dev-only
It is not good for there to be more than one version of derivable around at runtime.
bad things happen silently. I'm making ddom not explicitly dependent on derivable to
avoid silent bad things. module not found is lound and sensible. It also means the
versions aren't coupled. Yucky I know, but this is the best of several bad choices
I think.